govulncheck-action: support go-version-file
This change allows support for go-version-file in this action so that
consumers can reference a `go.mod` or `go.work` file rather than keeping
their action in sync with their go tooling. As actions/setup-go has
go-version overwrite go-version-file, this will do the same. If both are
provided, the go-version-input is used.

Fixes golang/go#61343