action.yml: make sure the latest Go version is default
Due to some confusion on the contract of one of the dependencies,
govulncheck-action would not by default try to download the newest go
version when one was present. This would sometimes even break the action
run due to a seemingly another issue with the dependency. This change
should fix both issues.

Fixes golang/go#69597

Change-Id: I51a21aad2daafc3984866051f91cf9339bb540e2
Reviewed-on: https://go-review.googlesource.com/c/govulncheck-action/+/616435
Reviewed-by: Maceo Thompson <maceothompson@google.com>
TryBot-Bypass: Zvonimir Pavlinovic <zpavlinovic@google.com>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>